Refactor: simplify ipfs dev/prod systems
